delete folder permission

+++ src/application/use-cases/delete-folder.ts
@@ -1,12 +1,19 @@
 import { DeleteFolderRepository } from '../../data/delete-folder-repository'
 import { DeleteFolder as IDeleteFolder, DeleteFolderData } from '../../domain/use-cases/delete-folder'
+import { UserHasFolderPermission } from '../../domain/use-cases/user-has-folder-permission'
+import { PermissionError } from '../errors/permission-error'
 
 export class DeleteFolder implements IDeleteFolder {
   constructor(
-    readonly deleteFolderRepository: DeleteFolderRepository
+    private readonly deleteFolderRepository: DeleteFolderRepository,
+    private readonly userHasFolderPermission: UserHasFolderPermission
   ) {}
 
   async exec(data: DeleteFolderData) {
+    const hasPermission = await this.userHasFolderPermission.exec({ userId: data.userId, folderId: data.id })
+    if (!hasPermission) {
+      throw new PermissionError()
+    }
     return await this.deleteFolderRepository.deleteFolder(data)
   }
 }
